requirements-dev.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
iniconfig==2.0.0
mysqlclient==2.2.5
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
